aider: Fix callback function to use `this.lastID` for newly inserted song ID.
# Aider chat conversation:

USER: The node server crashes when I try to add a song. This is the message: "POST /songs/add route hit
/Users/jackmazac/Documents/tunify_final/tunify/server/controllers/spotifyController.js:36
        res.status(201).json({ message: 'Song added successfully', songId: results.insertId });
                                                                                   ^

TypeError: Cannot read properties of undefined (reading 'insertId')
    at Statement.<anonymous> (/Users/jackmazac/Documents/tunify_final/tunify/server/controllers/spotifyController.js:36:84)
    at Statement.replacement (/Users/jackmazac/Documents/tunify_final/tunify/node_modules/sqlite3/lib/trace.js:25:27)
    at Statement.replacement (/Users/jackmazac/Documents/tunify_final/tunify/node_modules/sqlite3/lib/trace.js:25:27)

Node.js v18.17.0"

ASSISTANT: The error you're encountering is due to the `results` object not having an `insertId` property after inserting a new song into the database. This is because the `sqlite3` library's `run` method does not return a `results` object. Instead, the `this` context within the callback contains the `lastID` property, which is the ID of the last inserted row. I will provide a diff to fix this issue in the `spotifyController.js` file.

```diff
--- server/controllers/spotifyController.js
+++ server/controllers/spotifyController.js
@@ ... @@
-exports.addSong = (req, res) => {
+exports.addSong = (req, res) => {
     const { title, albumId, duration } = req.body; // Include other fields as necessary
     const query = "INSERT INTO Songs (title, albumId, duration, deleted) VALUES (?, ?, ?, 0)";
-    db.run(query, [title, albumId, duration], (err, results) => {
+    db.run(query, [title, albumId, duration], function(err) {
         if (err) {
             return res.status(500).json({ error: err.message });
         }
-        res.status(201).json({ message: 'Song added successfully', songId: results.insertId });
+        res.status(201).json({ message: 'Song added successfully', songId: this.lastID });
     });
 };
```

This change corrects the callback function to use the `this.lastID` property to get the ID of the newly inserted song and send it back in the response.