aider: Add Album and Duration columns to Songs table in init.sql.
# Aider chat conversation:

USER: Give me command to add artists to songs table in command line for sqlite3

ASSISTANT: To add artists to the `Songs` table in SQLite3 via the command line, you would use the `sqlite3` command to execute an `ALTER TABLE` statement. Here is the command you would run in your terminal:

```sh
sqlite3 path_to_your_database.db "ALTER TABLE Songs ADD COLUMN Artist VARCHAR(255);"
```

Replace `path_to_your_database.db` with the actual path to your SQLite database file.

This command will open the SQLite3 database and execute the SQL statement to add a new column named `Artist` of type `VARCHAR(255)` to the `Songs` table. If the `Songs` table already has an `Artist` column (as per the recent changes to the `init.sql`), this command should not be run again, as it would result in an error stating that the column already exists.

USER: Error adding song: Error adding song: SQLITE_ERROR: table Songs has no column named album fix same error for album and any other columns that need to be added

ASSISTANT: To fix the error regarding the missing `album` column in the `Songs` table, we need to update the `init.sql` file to include this column in the table schema. Additionally, we should review the `init.sql` file to ensure that all other necessary columns are present to match the expectations of the application code.

Here is the diff for the changes to the `init.sql` file to add the `album` column and any other missing columns to the `Songs` table:

```diff
--- database/init.sql
+++ database/init.sql
@@ ... @@
 CREATE TABLE Songs (
     SongID INTEGER PRIMARY KEY AUTOINCREMENT,
     AlbumID INT,
     Title VARCHAR(255) NOT NULL,
     Artist VARCHAR(255),
-    Duration TIME,
+    Duration TIME NOT NULL,
+    Album VARCHAR(255) NOT NULL,
     deleted INTEGER DEFAULT 0,
     FOREIGN KEY (AlbumID) REFERENCES Albums(AlbumID)
 );
```

This change adds the `Album` column to the `Songs` table and ensures that the `Duration` column is marked as `NOT NULL`. After applying this change, you will need to recreate the database or alter the existing table to add the new column. If you have existing data in the `Songs` table, you will need to consider how to migrate that data or handle the schema change. If the database is in development and can be reset, you can simply drop the `Songs` table and recreate it using the updated `init.sql` script.