aider: Add registration endpoint to userRoutes.js.
# Aider chat conversation:

USER: Create api endpoint for register

ASSISTANT: To create an API endpoint for registration, we need to modify the `userRoutes.js` file to include a new route that handles the registration logic. This route will use the `UserModel` to insert a new user into the database.

Since the `userRoutes.js` file is not currently in the list of *read-write* files, I cannot provide a diff for the changes. If you would like to proceed with this modification, please add `userRoutes.js` to the list of *read-write* files, and I will be able to assist you further.

USER: I added these *read-write* files: /Users/jackmazac/Documents/tunify_final/tunify/server/routes/userRoutes.js

ASSISTANT: To create the API endpoint for registration, we will add a new route to the `userRoutes.js` file. This route will handle POST requests to `/register` and will utilize the `UserController` to process the registration logic.

Here's the diff for adding the registration endpoint:

```diff
--- server/routes/userRoutes.js
+++ server/routes/userRoutes.js
@@ ... @@
 const UserController = require('../controllers/UserController');

+// Registration endpoint
+router.post('/register', UserController.register);
+
 router.get('/', UserController.getAllUsers);
 router.get('/:id', UserController.getUserById);
 router.post('/', UserController.addUser);
 router.delete('/:id', UserController.deleteUser);
 router.put('/:id', UserController.updateUser);

 module.exports = router;
```

This diff adds a new POST route for `/register` that will be handled by the `register` method in the `UserController`. The `UserController` will need to have a `register` method implemented that handles the registration logic, including validation, password hashing, and inserting the new user into the database. If the `UserController` does not already have a `register` method, it will need to be implemented.